destructure meta into individual components
