+++ src/pages/recipe/index.tsx
@@ -1,0 +1,10 @@
+import RecipeProvider from "../../context/RecipeContext";
+import Recipe from "./Recipe";
+
+export default function Root() {
+  return (
+    <RecipeProvider>
+      <Recipe />
+    </RecipeProvider>
+  );
+}
