+++ src/pages/recipe/recipe-categories/recipe-categories.module.css
@@ -1,16 +1,26 @@
 .container {
   display: flex;
-  overflow: scroll;
-  grid-column: 1/-1;
-  -ms-overflow-style: none; /* Internet Explorer 10+ */
-  scrollbar-width: none; /* Firefox */
-  margin-bottom: -1.5em;
-}
-.container::-webkit-scrollbar {
-  display: none; /* Safari and Chrome */
+  flex-direction: column;
+  gap: var(--section-gap);
 }
 
 .tags {
   display: flex;
   gap: 0.5em;
+  flex-wrap: wrap;
+  background-color: var(--card);
+  border-radius: var(--radius);
+  padding: var(--section-padding);
+}
+
+.button {
+  background-color: transparent;
+  border: none;
+  color: var(--text);
+  font-weight: 600;
+  font-size: var(--text-sm);
+  cursor: pointer;
+}
+.button:hover {
+  color: var(--primary);
 }
