--- src/pages/recipe/recipe-meta/recipe-meta.module.css
@@ -1,34 +1,0 @@
-.meta_lg {
-  display: grid;
-  grid-template-columns: 0.75fr 1fr;
-  grid-column: 1/-1;
-  gap: var(--layout-gap);
-}
-
-.right {
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-}
-
-.meta_sm {
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-}
-
-.title {
-  font-size: var(--text-2xl);
-  font-weight: 800;
-}
-
-.description {
-  font-weight: 450;
-}
-
-.image {
-  border-radius: var(--radius);
-  /* aspect-ratio: 1/1; */
-  object-fit: cover;
-  height: 100%;
-}
